Work-around for nested deletion
Cozy-desktop failed after doing `mkdir a/b/c/d ; rm -rf a` in a watched
folder. I think it's a race condition in PouchDB, but I didn't reproduce
it on a simpler test case. For the moment, this commit is a work-around
for this bug. But for the long-term, we should find the real bug.