Fix for images on write-only
When an image is uploaded on the cozy, its metadata are updated in the
CouchDB to include a thumbnail. So, for the write-only mode, it is
expected that the remote revision can sometimes be desynchronized with
the revision stored in the local PouchDB. This commit adds a way to
bypass this desynchronization and allow to update/remove the images.