Explain that node 4 is recommended
Node 0.10 and 0.12 will reach their end of life this year. So, it's time
to make node 4 the default. We need to test cozy-desktop with node 5 as
we will use it inside Electron for the GUI and Electron comes with node
5. Node 0.10 is still used a lot, in particular on Ubuntu and Debian, so
we should keep the tests for it too. But I've disabled the build for
node 0.12 to focus the support on the three other versions.