Use bash instead of maybe-missing rimraf in gui reinstall script
Using bash is definitely not the better way but it used to work,
while rimraf is always missing at this step of the release process,
forcing use to edit scripts by hand, not so nice when trying to
quickly release a fix.

This should be improved later by using the root node_modules/ for
dev tasks.