gui: Add synced dir to macOS favorites with python script
Apple removed the `sfltool add-item` subcommand in macOS 10.13 (High Sierra):
https://openradar.appspot.com/35722438

We were using it to add the sync folder to the Finder's favorite items on
previous macOS releases.

The `sfltool` CLI is actually a frontend to the `sharedfilelistd` XPC service.
The latter reads/writes from bplist files (e.g.
`~/Library/Application\ Support/com.apple.sharedfilelist`).

Talking to the XPC service was not a straightforward option.
Writing directly to those files could easily break. And the npm
packages we tried were losing data in the process.

A deprecated API exists to add favorite items:
https://developer.apple.com/documentation/coreservices/klssharedfilelistfavoriteitems

Having an embedded binary working on multiple macOS releases is not so easy,
so we used the built-in macOS python for this, since it includes bindings for
the relevant API.

We're not the only ones to use the deprecated API:
https://github.com/nextcloud/desktop/blob/master/src/common/utility_mac.cpp#L30
So hopefully we should be able to find a solution in case it is finally
removed in some future macOS release.