chore: Use an array instead of strings
As we use flow, we want to add some seriousness about our types.
By using the `...rest` parameter, we can assume we get an array
but we cannot know the types of its items.
By changing the interface, we tell flow we are awaiting a Array<string>
as the first parameter and a Function as the second.