Add a script to visualize logs with gource (#1336)
* Add a script to visualize logs with gource

Gource is a tool made for visualizing a repository (git, hg, svn, cvs)
evolution in time. It also accepts log in a custom format. I wanted to
try to visualize the logs from cozy-desktop with it, as a friday hack.
There is a script to do that. It looks beautiful (thanks to gource), but
I don't think it will really help us for debugging cozy-desktop. It's
here, just for the fun :-)

Gource: https://gource.io/

* fixup! Add a script to visualize logs with gource