test/scenarios: Capture move dir + update subfile locally
Add delay between actions to prevent events squashing.

Captured on GNU/Linux and Windows only.
The macOS capture is the same as GNU/Linux.

On Windows, for some reason the change event seems to be swallowed.
But since we compute the md5sum and it is different, the file is
effectively updated in the end.