core/local: Replace change.fromEvent() with dedicated builders
Flow typechecking is terribly inaccurate in `local/analysis` & `local/change`.
The `change.fromEvent()` function in particular is messing up flow in unexpected ways.
For example, it prevents using the `?: ...` syntax (vs `: ?...`) in a few type definitions.

Replacing it with dedicated builders is a first necessary step to regain sanity.

The builders currently include some duplication.
But this should be easy to DRY later, once the code is more reliable.

Also the `LocalFileAddition` type is currently lying: `md5sum` is missing when `wip` is `true`.
Fixing it will require quite more work, hence the temporary `delete md5sum` hack when a file
is added then unlinked.