core/local/analysis: Add missing break statements & drop useless ones
In two branches, a `case` statement was missing, which means the next branch
was also executed. But since we can only have one change per path in
changesByPath, it didn't cause any harm.

In one branch, the `case` statement was redundant with the final one.

So this commit should not change the existing behavior in any way.