core: Separate Config class & module
A module will allow us to export logic independent from config file as
plain functions

It makes the following not so nice, though:

1. `new Config.Config()`
2. `Config.Config.safeLoad()`
3. `Config.Config` type annotations

Points 1 and 2 can easily be improved (see subsequent commits)

Regarding point 3, there's not so much we can do...
Aliasing module as `config` would conflict with existing vars.
Because JS declared vars and imported modules are in the same scope.
Importing type as `Config` is not possible either.
Because flow types are in the same scope as everything else.
Suggestions are welcome :)