core/local/steps/incomplete_fixer: Handle A→B→C→D
When we move a file twice in a row, we would not complete the event
correctly and end up with C→D instead of A→D. Therefore, we wouldn't
be able to merge the move.
To handle this case, we go over incomplete events to complete other
incomplete events (e.g. B→C will complete A→B) but in such cases, the
rebuilt event will be marked as incomplete as well and will be offered
for completion later on.

A side effect of this fix is to drop useless complete events (e.g.
deleted event for A when we completed A→B with deleted B).