core: Pass all Merge args to IdConflict
Both `side` & `doc` are part of the same `change`.
It is this `change` that conflicts with `existingDoc`.
Hence the `Change` type introduced only in `IdConflict`
for now.
Soon we'll use `was` to prevent identity conflict in
case of an identical renaming.

To make the transition easier, `sideName` and `newDoc`
were renamed to `side` and `doc` to better match the
existing wording in `Merge` and make it easier to wrap them
in a `change` object.

We can still rename everything at once later.
Also it could also be a good idea to pass `Change` objects
to all `Merge` methods at some point.