core/local: Introduce ./chokidar namespace
To make it easier to see which part of the code is specific to our
`ChokidarWatcher`.

Rename `./chokidar_*` to `./chokidar/*`. Move `./analysis` and
`./event_buffer` there too.

Rename `./event` and `./change` to ~./chokidar/local_*`: they were
initially not supposed to be specific, but they ended up being so. Big
design up front failing as usual. Prefixing with `local_` to match their
type names, to prevent confusion with `./chokidar/event`.

This commit is only about moving files and fixing import paths, fixture
paths and test descriptions.