yarn: Remove mocha direct dev dependency
It is already a dependency of electron-mocha.
So we don't install two different versions of it.
Although it forces us to upgrade to 6.x, but CI is green anyway.

Also eslint complains about the extraneous mocha requirement in
`test/support/istanbul_reporter`, but we don't want to require
`electron-mocha` instead, neither do we want to reinstall `mocha`.
So let's ignore it for now.