ci/travis: Stop sending PR builds to Bintray (#1689)
We haven't really made use of the binaries built for each PR and
pushed to Bintray in a long time.
In the meantime, our Bintray account has been locked because we've
reached our storage limit. Since unlocking the account requires to
contact their support even after we've removed some of the stored
assets, we have postponed it until we felt it was necessary. We
haven't done it yet.
Those push take time and fail from time to time, requiring us to
restart the build job.
As a temporary solution to those problems, we'll stop sending the
binaries until we have a need for them again.