core/merge: Transform child move if file was created on other s… (#1637)
If a file is added to a directory on one side while the directory is
moved on the other side, we should not try to apply the move on the
child file but transform it instead so it is interpreted as a file
addition at the new location (i.e. in the directory's new path).
This is allowing us to merge both changes during the initial scan
where the client applies changes coming from both sides before
synchronizing them.