gui/tray: Add complete context menu on all Linux (#1714)
Starting with Electron 3 (still true with Electron 4 and potentially
with Electron 5 as well), whenever libappindicator-gtk3 is installed
on the client's computer, clicks on the systray icon won't trigger
click events.

We use those events to display the app's window and have currently no
other ways to display it on GNOME systems.
To make sure all users will have at least one way to display the app's
window, we add the complete context menu (i.e. not only the
« Quit Application » menu item on right click but also the
« Show Application » menu item, both on left click) on all Linux
systems.
If the libappindicator-gtk3 package is not installed (i.e. it is a
dependency of the Chrome Web browser), the left click will still
display the app's window.

On a side note, it was our intention to display part of the context
menu on Mac OSx (the « Quit Application » item). While this item is
actually displayed on right-click, we were never adding a context menu
to the systray icon because of an inverted condition.
This is now made clear what the behavior is (i.e. no context menu).