dist: Disable sandbox on kernels not supporting it (#1762)
Electron packages a Chromium executable within each app to render the
application's graphical content and starting with Electron v5, the
Chromium sandbox is activated by default.
On Linux, recent versions of Chromium use a kernel feature to run this
sandbox: unprivileged user namespaces with clone. This feature is
unfortunately disabled by default on Debian issued kernels as they
consider it is not ready for production and increases the attack
surface of the operating system (see
https://lwn.net/Articles/673597/).

Chromium has another option to run sandboxes: using a root owned
binary with the SUID bit to access the privileged user namespaces.
This is a separate binary which is shipped with Electron applications.

Our problem is that the SUID binary is packaged within our AppImage
and is therefore not owned by root so it doesn't have acces to the
user namespaces operations.
Our only way is then to disable the Chromium sandbox via the
`--no-sandbox` flag.

This commit adds a bash launcher setting the flag only when the kernel
does not support unprivileged user namespaces so users of upstream
kernels can benefit of the sandbox while users of Debian issued
kernels can use the application even with decreased security (this is
very limited since we only load web pages from cozy.io during the
onboarding and then load only local html files).