doc: Update supported macOS versions (#1767)
It has been reported to us that the application does not work on
macOS 10.9 Mavericks anymore. Since we don't have any device to test
this, Apple does not officially support this version anymore and the
Chrome documentation states that only macOS 10.10+ is supported, we've
decided to show we don't support this version at all and that we don't
expect Cozy Desktop to work on it anymore.

We also take the opportunity to add the last 2 versions of macOS and
state that we only actively support those 2 while only the latest
version is actively tested since we have a device to test it.