gui: Debounce sync status updates display (#1771)
With the Atom Watcher, changes are handled one at a time and this can
lead to status "blinking" in the systray and the app's status bar
between the synchronization state and other states like up-to-date or
change preparation (i.e. the change is being processed by the watcher
or being merged).
This can be off-putting for users who are not aware of this
behind-the-scene mechanism and might be feeling the application is
acting erratically.

By simply debouncing status updates display in either in the systray
or the status bar when this is not a long-term status (i.e. offline,
error, syncing, online after offline), we can give a sense of
continuity and that the process is going smoothly.