core/pouch: Pick correct previous revision (#1773)
We changed the arguments of `Pouch#getPreviousRev()` to get passed a
revision diff count so we could introduce the `sides.target` on
documents.
This count is meant to select the count'th previous revision of the
specified document. To do so, we fetch all revision ids of the
document (i.e. the hash part of each revision) and use the count to
pick the id at the right index. We thought those were ordered by
ascending short revision order while they're actually ordered by
descending short revision order (i.e. the latest revision id is the
first of array).

This means that we were not selecting the right revision in Sync when
trying to get the latest known remote revision to propagate a file
update to the Cozy. This would cause conflict errors and prevent the
propagation.