gui/elm: Disable manual sync after first click (#1790)
When clicking on the manual synchronization button in the Settings
pane, we send a message to the main thread to restart in the Sync
flow.
However, until we receive the first status update back from the
synchronization process, the button stays enabled and we can
mistakenly request more manual synchronizations (e.g. by
double/triple/... clicking the button).

This can potentially lead to mistakes during the triggered initial
scan and should be avoided.

We introduce a new field in the Settings model that keeps track of
manual synchronizations. It gets updated when we click the button and
when we receive an up-to-date status back from the synchronization
process.

Also, instead of duplicating the Tray's status into the Settings 
model, we can just pass it to the view as argument.