core: Mark local versions of Notes as read-only (#1791)
Since the introduction of the Cozy Notes application, we've started
synchronizing their markdown version on users' desktops.

The problem is, those version are not meant to be modified as they're
just a markdown export of the actual Note and modifications made on
those files with any application other than Cozy Notes won't be
reflected on the Note stored in the Cozy.

To avoid situations where users modify those files and lose content or
just don't find it on their Cozy, we've decided to use file system
permissions to show those files should not be modified.
Users can still change the permissions but we find this is a good
warning while we add the capacity to open Notes from the file system
(i.e. we'll launch the Cozy Notes application for that).

Warning: This behavior is based on the premise that Cozy Notes
related files will have the `text/vnd.cozy.note+markdown` MIME type.
This is not the case yet!
But we're relying on the idea that users will run a version of the
Desktop app with this behavior before documents are migrated on their
Cozy to have the correct MIME type. Cozy Desktop will detect the
change and apply the new permissions on the local file system.
Any other change on Notes will also trigger a permissions change.