core/sync: Don't setup Pouch during manual sync (#1794)
The PouchDB views and migrations are setup during the start of the
Sync module. This seemed appropriate as this would make sure the DB
was ready before launching the watchers.

However, we've since reused the `Sync#start()` method to run manual
synchronizations. This means that views are created and migrations run
every time users click the manual synchronization button.
Even if these should not have any effects after the first run, it
seems risky and unnecessary to run them at this point.

It seems reasonable to do these actions during the application setup
which will happen before the synchronization is started.