tools: Associate `.cozy-note` files with Desktop (#1801)
Create a file association on the user's computer via the
`electron-builder` configuration.
This will create the `text/vnd.cozy.note+markdown` MIME type on macOS
and Windows and will set Cozy Desktop as the default application to
open files with the `cozy-note` extension.

On Linux, since we don't install the application but hand an
executable, we need to create the MIME type by hand by putting a
declaration file in the local MIME types database.
The association will only work if a `desktop` file with the right MIME
type association points to the Desktop app. This is done automatically
if using `appimaged`.