core/local/chokidar: Reuse non-NFC path on HFS+ (#1803)
When synchronizing a remotely added document with a path containing
utf-8 characters on an HFS+ file system, we have to take care of the
re-normalization done by the file system itself (i.e. it renames the
document with `NFD` normalized path.
To avoid treating those renames as normal movements (invalid since
they give the same PouchDB document id via the `NFD` normalization
done to get the id of documents on macOS), we check if an existing
document with the same equivalent path exists in PouchDB and if so, we
normalize our document path with `NFC` to make sure it won't trigger a
rename action if it is strictly equal to the existing path.

However, this only works if the entire existing PouchDB document's
path is normalized with `NFC`.
We have encountered cases where the remote document path had its name
normalized with `NFC` (e.g. created on the Cozy by a connector) while
its ancestors were normalized with `NFD` (e.g. created locally and
thus synchronized with this norm).

We propose to re-use the existing document's ancestors part so don't
change its normalization and only apply the `NFC` normalization to the
name of the document.

The whole point of those re-normalizations are to avoid detecting
false renamings when it's only the HFS+ normalization process that
kicks in.