gui/main: Setup App only when we have a config (#1814)
New users or users who disconnect their client don't have an App
config. This prevents us from running the `App.setup()` method since it
will raise an error for invalid configs (like an empty one).

The config is built during the On-boarding process so we need to
postpone the setup of the App after the user has gone through that
process.

This means notes can't be opened if the client has not been connected
before.