core/sync: Catch sync errors when requesting stop (#1820)
Errors raised from the sync process, the local watcher or the remote
watcher reject the `started` and `running` Promises that we use to
detect them from the main process.

However, the `Sync.stop()` method awaits them to make sure we don't
stop the process before all its modules have been initialized (to
avoid running them more than once in case we're restarting the sync
process).
This means that if those Promises are rejected, the `stop` method
itself will raise an error and the sync process won't be stopped.
This is an issue when we try to run some actions after stopping the
sync process (e.g. disconnecting the client after the OAuth client has
been revoked) since errors during the `stop` call will prevent further
actions to be run via calls to `then.