gui/main: Change reconnection button label (#1824)
When we detect the OAuth client has been revoked, we offer the user to
fully disconnect Cozy Desktop from her Cozy and we restart the
application.
Doing so, the user is presented with the on-boarding process again.

To better reflect this behavior, we changed the button label to
"Reconnect". We believe users that disconnect their Cozy mostly do so
to reinitialize their app or by mistake.