core/local/atom: Default event kind to file (#1826)
In some situations, the stats returned by the file system for a given
path don't contain a type (i.e. file, directory, symbolic link…).
This means that we won't know what kind of action to apply on
documents generated by events for this path.

We decided a while ago to default the event kind for such paths to
`file` in the hope we could reduce the amount of situations where the
kind is unknown or wrong.

However, some users already have Pouch documents without doctypes so
we have to use a default value when inferring the event kind from it
(i.e. during the initial diff when we created `deleted` events for
documents we couldn't match with a file or directory on the file
system).
We'll use the same default type of `file`.