sync: Propagate move with incompatibilites to Cozy (#1837)
If a document with incompatibilities is moved locally (i.e. via one of
its ancestors) we should propagate the movement to the remote Cozy and
not transform it into an addition because the document exists on the
Cozy and not on the local file system (we'd end up with an error
because the file could not be read and the Cozy would reject the
addition anyway).

We'll still transform the remote movements of documents with platform
incompatibilities into additions since the document does not exist on
the local filesystem and therefore could not be moved.