core/local: No errors on missing trashed file/dir (#1839)
When trying to put a remote document to the Cozy trash, if it's
missing, we skip the error since there's nothing we can do about it.
We should do the same when trying to trash a local file.

There's a slight difference when trashing a folder since we can get
the `ENOENT` system error (i.e. missing document) if the document at
the given path is a file instead of a folder.