build: Unpack native extensions (#1846)
As stipulated in the Electron documentation about application
packaging, files are packaged by default into an `asar` archive to
speed up some require calls and prevent path name length issues on
Windows.
However, some Node API calls do not work with packed files and need to
unpack them first. This is the case for native extensions (i.e.
`.node` files) and this operation can trigger anti-viruses, thus
throwing exceptions (see https://www.electronjs.org/docs/tutorial/application-packaging#adding-unpacked-files-to-asar-archives).

We believe some of our users can be experiencing those issues and as a
workaround, we propose to not pack (or unpack in Electron terms) the
native extensions when building the application.