gui/main: Instanciate sync only once (#1847)
The `main` module is responsible for instantiating the synchronization
process and thus all its sub-modules (i.e. watchers and sync mainly)
once the proxy and config and set.

With the different states in which the app can be started (i.e. not
connected to a Cozy, waiting for an update, already connected, opening
a note…), we have different paths leading to an instantiation.
However, we should not ever instantiate the app more than once or we
could see multiple instances of the local watcher running and maybe
different processes accessing the PouchDB database file.

Unfortunately, we were instantiating the sync twice in case the client
was already connected to a Cozy and running in normal, synchronization
mode.