.cozyignore: Ignore all Microsoft Office tmp files (#1849)
We want to make sure all temporary files created by a Microsoft Office
software are ignored to avoid synchronizing them and later detect a
movement instead of a file update when the edited file is saved.

We were ignoring temporary files for `.doc*`, `.xls*` and `.ppt*`
documents only (i.e. and not the open document types).