core/remote/watcher: Missing parents don't block (#1845)
We decided to block the synchronization when the client could not merge
a change because its parent folder could not be found in PouchDB (i.e.
`MergeMissingParentMetadataError`).

However, some of those situations can resolve themselves by skipping
those changes and trying again later, after the other side changes
have been propagated (non exhaustive list):
- a folder was renamed locally while some content was added in it on
  the remote Cozy (the content ends up in the renamed folder)
- a folder was deleted locally while some content was added in it on
  the remote Cozy (the folder and all its content are trashed)

The decision to block the synchronization was made at a time when we
didn't have great confidence in the client and we preferred to have
users contact us whenever they encountered an issue so that we could
analyze their logs as soon as possible.
Unfortunately, this generates a lot of unsolvable errors for the user,
without any information about its source and at the worst time (when
we fail to merge a child change and not when we have an hypothetical
error when merging the parent change). In this situation, it is very
likely that users won't contact us or too late to have access to the
log files we specifically wanted to get.

So we decided to revisit that decision and stop blocking
synchronizations for `MergeMissingParentMetadataError`s.
Errors will keep showing up in Sentry and users who finally find
discrepancies between their local synchronization folder and their
Cozy will still contact us anyway.
We hope that most sources of this error can be fixed by another round
of synchronization.