core/merge: Mark documents for deletion (#1836)
Until now, when we detected the deletion of a document either on the
local file system or the remote Cozy, we would merge it by deleting
it from the PouchDB database.
Then, The Sync module would detect that change via the PouchDB changes
feed and propagate it.

However, deleted documents in PouchDB cannot be queried by `get()`.
Also, when multiple modifications are made to the same document (i.e.
saved with the same `_id` attribute), the changes feed will return
only the latest version of the document.

Those 2 "restrictions" make merging some changes harder. For example:
- the remote addition of files or folders into a locally deleted
  folder (i.e. the parent folder can't be queried so it's "missing")
- the deletion and creation of a file at the same path/_id (i.e. we
  lose the _rev of the deleted document and can't propagate it as an
  update)

The solution we found to this problem is to mark documents for
deletion during the Merge step so they can still be queried until the
deletion has been propagated. This way we can decide to cancel the
deletion or take any other appropriate action to merge other changes
that could not be merged (or merged in a way that would result in an
incorrect or impossible propagation).

This commit introduces the `deleted` marker and does not make use of
it to solve our problems yet. Its goal is to achieve feature parity.