core/remote/watcher: Restored docs are always added (#1867)
Since we introduced the deletion marker, documents that were trashed
on the remote Cozy and restored before the deletion could be applied
on the local filesystem (i.e. the new version of the PouchDB record
has been merged with the marker but the document still exists) would
be considered as moved instead of being considered as added.

We missed a condition in the remote watcher where we needed to check
for the presence of the marker because we were checking for the
presence of a `trashed` attribute.
However, this attribute does not have the same meaning and would
actually not be present on any record in the PouchDB database since
it's a remote CouchDB attribute set on trashed documents.