core/remote/watcher: Skip sequences missing parent (#1884)
We've stopped blocking the synchronization process when encountering
only `MergeMissingParentError`s in a single remote watcher changes
batch in f5c05d5c32a0b439d9456825ef64b8ddb41d9c91.

However, this commit did not change the way we manage the remote
sequence when encountering those errors. This means that even though
the synchronization does not stop, we do not mark the merged changes
as processed and we will fetch them and merge them again at the next
remote watcher poll.
In some cases, we can see missing parent errors that will never get
resolved and we'll keep fetching and merging the same changes over
and over again. Besides the inefficiency of doing the same work every
minute, we could mess up the local PouchDB records by merging a change
more than once, especially if the document has changed on the local
filesystem and the record was updated accordingly.

Also, we believe that most of those errors are now coming from stack
errors. Those errors are being fixed and so we should not see their
repercussions anymore very soon.
In this situation, we prefer not to do any extra work to handle those
changes that could not be merged.

This is why we've decided to mark changes that encountered missing
parent errors as processed anyway by updating the remote sequence in
our PouchDB database.
If the documents are changed again on the remote Cozy, the remote
watcher will try again to merge them, as any other change.