core/merge: Remove Note metadata on conflicts (#1895)
When we detect a local Cozy Note export modification and create a
conflict, we want to mark the new markdown file as such and not a real
Cozy Note.

We won't change the file's extension in case the file is still open in
the user's editor and so we won't change the document's mime type
either. However, we'll make sure the Cozy Note's content and schema,
stored in the existing record's `metadata` attribute, won't be part of 
the new record since the actual file's content has changed and won't 
be re-imported as a Note.

Without these attributes, Drive Web should be able to tell that the 
file needs to be displayed as a markdown file.